Split widget: insight row
